Fix to the bug of `getRecordsByPrefix`
* If there are more than one matching records in the same log file,
and some of them are already flushed to the table section, the
remaining unflushed records will be ignored even though their prefix
matches.

* Fixed the logic in memtable.
