+++ scripts/runtests.sh
@@ -23,6 +23,7 @@ $VALGRIND ./tests/memtable_test --abort-on-failure
 $VALGRIND ./tests/table_lookup_booster_test --abort-on-failure
 
 $VALGRIND ./tests/basic_op_test --abort-on-failure
+$VALGRIND ./tests/nearest_search_test --abort-on-failure
 $VALGRIND ./tests/seq_itr_test --abort-on-failure
 $VALGRIND ./tests/key_itr_test --abort-on-failure
 $VALGRIND ./tests/snapshot_test --abort-on-failure
