+++ tests/CMakeLists.txt
@@ -52,6 +52,11 @@ add_executable(basic_op_test ${BASIC_OP_TEST})
 target_link_libraries(basic_op_test ${JUNGLE_TEST_DEPS})
 add_dependencies(basic_op_test static_lib)
 
+set(NEAREST_SEARCH_TEST ${TEST_DIR}/jungle/nearest_search_test.cc)
+add_executable(nearest_search_test ${NEAREST_SEARCH_TEST})
+target_link_libraries(nearest_search_test ${JUNGLE_TEST_DEPS})
+add_dependencies(nearest_search_test static_lib)
+
 set(BUILDER_TEST ${TEST_DIR}/jungle/builder_test.cc)
 add_executable(builder_test ${BUILDER_TEST})
 target_link_libraries(builder_test ${JUNGLE_TEST_DEPS})
